version 1.3.8 fix printed expiration time
If the planned absolute expiration time was less than the default
absolute expiration time _from now_, then it would only print when the
player's claims should expire according to the default expiration
instead of accounting for any added delays.

I.e. if a player's expiration date is 5 days in the future, with the
default expiration being 3 days, the command will incorrectly state that
the player's claims should be pending expiration after the third day.

- Also prints extra details when canceling a claim expiration in console